refactor: split agentcore browser internals
